trocr/requirements.txt
torch
transformers
datasets
evaluate
Pillow
matplotlib
